requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
# websocket-client>=1.6.0  # Commented out - see signalrcore section below for version
aiohttp>=3.9.0
pyyaml>=6.0.1

# Optional: SignalR client for real-time updates (not required - REST polling works fine)
# NOTE: signalrcore has dependency conflicts with websocket-client>=1.6.0
# If you want SignalR support, install manually: pip install signalrcore==0.9.5 websocket-client==1.0.0
# The system works fine without SignalR using REST polling
# signalrcore>=0.9.0  # Uncomment and adjust websocket-client version if needed

#installed
signalrcore==0.9.5 
websocket-client==1.0.0

# Data & storage
sqlalchemy>=2.0.0
pandas>=2.0.0
numpy>=1.24.0

# API & web server
fastapi>=0.104.0
uvicorn>=0.24.0
prometheus-client>=0.19.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Code quality
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0

